Make generic structs parse
