+++ crates/parser/src/grammar.rs
@@ -1,5 +1,6 @@
 pub mod contracts;
 pub mod expressions;
 pub mod functions;
+pub mod generics;
 pub mod module;
 pub mod types;
