+++ crates/tests/fixtures/files/generic_functions_return_t.fe
@@ -1,0 +1,8 @@
+struct Foo<T> {
+  pub val: T 
+}
+
+#test
+fn test() {
+  let x: Foo<u32> = Foo ( val: 1 );
+}
